requirements.txt
mcp>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0
